+++ pg-ephemeral/tests/backtrace.rs
@@ -1,4 +1,4 @@
-const RUST_BACKTRACE: cmd_proc::EnvVariableName<'static> =
+const RUST_BACKTRACE: cmd_proc::EnvVariableName =
     cmd_proc::EnvVariableName::from_static_or_panic("RUST_BACKTRACE");
 
 #[tokio::test]